+++ src/collective/cover/config.py
@@ -25,3 +25,9 @@ DEFAULT_SEARCHABLE_CONTENT_TYPES = [
     'Link',
     'News Item'
 ]
+
+# z3c.form.widget.SequenceWidget used for css_class field
+# has a default null value of '--NOVALUE--'
+# In case that no value was set to that field
+# we need to detect it and use it
+DEFAULT_SEQUENCEWIDGET_VALUE = '--NOVALUE--'